feat: 아키텍처 구조 설계
